rename StanzaSuperClass -> MetaStanza
