+++ lib/MetaStanza.ts
